services/image_recognition_service/requirements.txt
fastapi;
uvicorn;
pytesseract;
Pillow;
python-multipart;